fix(client): message broker page
